+++ src/app/(root)/(routes)/(home)/page.tsx
@@ -1,7 +1,7 @@
 'use client'
 
-import { useEffect, useRef } from 'react'
 import CardGridTemplate from '@/components/templates/CardGridTemplate'
+import { useInfiniteScroll } from '@/hooks/useInfiniteScroll'
 import useGetAllPosts from '@/queries/channel'
 
 const CHANNEL_ID = process.env.NEXT_PUBLIC_CHANNEL_ID as string
@@ -9,24 +9,7 @@ const CHANNEL_ID = process.env.NEXT_PUBLIC_CHANNEL_ID as string
 export default function Home() {
   const { data, fetchNextPage, hasNextPage } = useGetAllPosts(CHANNEL_ID)
 
-  const observerElem = useRef(null)
-
-  useEffect(() => {
-    const element = observerElem.current
-    const option = { threshold: 0.5 }
-
-    const io = new IntersectionObserver((entries) => {
-      const [target] = entries
-      if (target.isIntersecting) {
-        fetchNextPage()
-      }
-    }, option)
-
-    if (hasNextPage && element) {
-      io.observe(element)
-      return () => io.unobserve(element)
-    }
-  }, [fetchNextPage, hasNextPage])
+  const { observerElem } = useInfiniteScroll({ fetchNextPage, hasNextPage })
 
   return <CardGridTemplate postDatas={data?.pages.flat()} ref={observerElem} />
 }
